fix image domains name
