+++ src/components/shared/Navbar.jsx
@@ -1,0 +1,65 @@
+import Image from "next/image";
+import React from "react";
+import { CiHeart, CiBellOn } from "react-icons/ci";
+
+const Navbar = () => {
+  return (
+    <div className="navbar bg-base-100">
+      <div className="flex-1">
+        <a className="btn btn-ghost text-xl">daisyUI</a>
+
+        <button className="btn btn-sm">
+          <CiHeart />
+        </button>
+
+        <button className="btn btn-sm">
+          <CiBellOn />
+        </button>
+        <button className="btn btn-sm">Button</button>
+        <button className="btn btn-sm">Button</button>
+      </div>
+      <div className="flex-none gap-2">
+        <div className="form-control">
+          <input
+            type="text"
+            placeholder="Search"
+            className="input input-bordered w-24 md:w-auto"
+          />
+        </div>
+        <div className="dropdown dropdown-end">
+          <div
+            tabIndex="0"
+            role="button"
+            className="btn btn-ghost btn-circle avatar"
+          >
+            <div className="w-10 rounded-full">
+              <Image width={50} height={50} 
+                alt="Tailwind CSS Navbar component"
+                src="https://daisyui.com/images/stock/photo-1534528741775-53994a69daeb.jpg"
+              />
+            </div>
+          </div>
+          <ul
+            tabIndex="0"
+            className="mt-3 z-[1] p-2 shadow menu menu-sm dropdown-content bg-base-100 rounded-box w-52"
+          >
+            <li>
+              <a className="justify-between">
+                Profile
+                <span className="badge">New</span>
+              </a>
+            </li>
+            <li>
+              <a>Settings</a>
+            </li>
+            <li>
+              <a>Logout</a>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Navbar;
